requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
Django==4.2.3
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.3
numpy==1.25.1
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
xlrd==2.0.1
